Move `ServerRoute` to `SharedModels`
