+++ Sources/ApiClient/Client.swift
@@ -1,7 +1,6 @@
 import Combine
 import ComposableArchitecture
 import Foundation
-@_exported import ServerRoutes
 import SharedModels
 
 public struct ApiClient {
