+++ Sources/DemoMiddleware/DemoMiddleware.swift
@@ -3,7 +3,6 @@ import DictionaryClient
 import Either
 import HttpPipeline
 import Prelude
-import ServerRoutes
 import SharedModels
 
 public struct SubmitDemoGameRequest {
