+++ Sources/ServerRouter/Router.swift
@@ -1,7 +1,6 @@
 import ApplicativeRouter
 import Foundation
 import Prelude
-import ServerRoutes
 import SharedModels
 import Tagged
