+++ Sources/ShareGameMiddleware/ShareGameMiddleware.swift
@@ -6,7 +6,6 @@ import Foundation
 import HttpPipeline
 import Prelude
 import ServerRouter
-import ServerRoutes
 import SharedModels
 
 public struct ShareGameRequest {
