+++ Sources/SiteMiddleware/ApiMiddleware.swift
@@ -14,7 +14,6 @@ import Prelude
 import PushMiddleware
 import ServerConfigMiddleware
 import ServerRouter
-import ServerRoutes
 import ShareGameMiddleware
 import SharedModels
 import SnsClient
