+++ Sources/SiteMiddleware/SiteMiddleware.swift
@@ -7,7 +7,6 @@ import MiddlewareHelpers
 import Prelude
 import ServerConfig
 import ServerRouter
-import ServerRoutes
 import ShareGameMiddleware
 import Tagged
