+++ Tests/DailyChallengeMiddlewareTests/DailyChallengeMiddlewareTests.swift
@@ -9,7 +9,6 @@ import HttpPipeline
 import HttpPipelineTestSupport
 import MailgunClient
 import Overture
-import ServerRoutes
 import SharedModels
 import SnapshotTesting
 import XCTest
